+++ inst/NEWS.md
@@ -1,3 +1,3 @@
 # table.express 0.2.0.9000
 
-- TBD
+- `distinct` gained `.keep_all` and `.keep_old` parameters.